+++ gentity-core/src/main/java/com/github/gentity/core/model/dbs/DbsDatabaseModel.java
@@ -15,6 +15,7 @@
  */
 package com.github.gentity.core.model.dbs;
 
+import com.github.gentity.core.Exclusions;
 import com.github.dbsjpagen.dbsmodel.ProjectDto;
 import com.github.gentity.core.model.DatabaseModel;
 import com.github.gentity.core.model.SequenceModel;
